chore: reorganize and cleanup files
