+++ core/lib/cart/index.ts
@@ -3,10 +3,10 @@
 import { unstable_expireTag } from 'next/cache';
 
 import { auth, getAnonymousSession, updateAnonymousSession, updateSession } from '~/auth';
-import { addCartLineItem, AddCartLineItemsInput } from '~/client/mutations/add-cart-line-item';
-import { createCart, CreateCartInput } from '~/client/mutations/create-cart';
-import { getCart } from '~/client/queries/get-cart';
 import { TAGS } from '~/client/tags';
+import { addCartLineItem, AddCartLineItemsInput } from '~/lib/cart/add-cart-line-item';
+import { createCart, CreateCartInput } from '~/lib/cart/create-cart';
+import { validateCartId } from '~/lib/cart/validate-cart';
 
 import { MissingCartError } from './error';
 
@@ -50,7 +50,7 @@ export async function addToOrCreateCart(
   data: CreateCartInput | AddCartLineItemsInput['data'],
 ): Promise<void> {
   const cartId = await getCartId();
-  const cart = await getCart(cartId);
+  const cart = await validateCartId(cartId);
 
   if (cart) {
     const response = await addCartLineItem(cart.entityId, data);
