+++ core/vibes/soul/primitives/navigation/index.tsx
@@ -27,8 +27,8 @@ import { Logo } from '@/vibes/soul/primitives/logo';
 import { Price } from '@/vibes/soul/primitives/price-label';
 import { ProductCard } from '@/vibes/soul/primitives/product-card';
 import { Link } from '~/components/link';
-import { useSearch } from '~/context/search-context';
 import { usePathname, useRouter } from '~/i18n/routing';
+import { useSearch } from '~/lib/search';
 
 interface Link {
   label: string;
