+++ core/vibes/soul/sections/not-found/index.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from '@/vibes/soul/primitives/button';
 import { SectionLayout } from '@/vibes/soul/sections/section-layout';
-import { useSearch } from '~/context/search-context';
+import { useSearch } from '~/lib/search';
 
 export interface NotFoundProps {
   title?: string;
